test: split default role integration coverage
